refactor ~ restructure src code (sequester impl details)
