+++ src/lib.rs
@@ -26,19 +26,10 @@ To see specific usage details, look at the `uname` utility linked above as it ma
 use of every feature.
 */
 
-pub use self::sys::*;
-
 use std::borrow::Cow;
 
-#[cfg(unix)]
-#[path = "unix.rs"]
-mod sys;
-#[cfg(windows)]
-#[path = "windows.rs"]
-mod sys;
-#[cfg(not(any(unix, windows)))]
-#[path = "unknown.rs"]
-mod sys;
+mod lib_impl;
+pub use lib_impl::*;
 
 /// `Uname` is meant for types that can provide information relevant to `uname`.
 pub trait Uname {
@@ -60,25 +51,3 @@ pub trait Uname {
     /// The name of the current OS.
     fn osname(&self) -> Cow<str>;
 }
-
-// private platform-specific HOST_OS_NAME * ref: [`uname` info](https://en.wikipedia.org/wiki/Uname)
-#[cfg(all(target_os = "linux", any(target_env = "gnu", target_env = "")))]
-const HOST_OS_NAME: &str = "GNU/Linux";
-#[cfg(all(target_os = "linux", not(any(target_env = "gnu", target_env = ""))))]
-const HOST_OS_NAME: &str = "Linux";
-#[cfg(target_os = "android")]
-const HOST_OS_NAME: &str = "Android";
-#[cfg(target_os = "windows")]
-pub const HOST_OS_NAME: &str = "MS/Windows"; // prior art == `busybox`
-#[cfg(target_os = "freebsd")]
-const HOST_OS_NAME: &str = "FreeBSD";
-#[cfg(target_os = "netbsd")]
-const HOST_OS_NAME: &str = "NetBSD";
-#[cfg(target_os = "openbsd")]
-const HOST_OS_NAME: &str = "OpenBSD";
-#[cfg(target_vendor = "apple")]
-const HOST_OS_NAME: &str = "Darwin";
-#[cfg(target_os = "fuchsia")]
-const HOST_OS_NAME: &str = "Fuchsia";
-#[cfg(target_os = "redox")]
-const HOST_OS_NAME: &str = "Redox";
